refactor: StudyMember에서 StudyMemberUIModel로 이름 변경
