+++ android/app/src/main/java/com/created/team201/presentation/studyDetail/adapter/StudyPeopleViewHolder.kt
@@ -4,13 +4,13 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.created.team201.databinding.ItemStudyDetailStudyPeopleBinding
-import com.created.team201.presentation.studyDetail.model.StudyMember
+import com.created.team201.presentation.studyDetail.model.StudyMemberUIModel
 
 class StudyPeopleViewHolder(val binding: ItemStudyDetailStudyPeopleBinding) :
     RecyclerView.ViewHolder(binding.root) {
 
-    fun bind(studyMember: StudyMember) {
-        binding.studyMember = studyMember
+    fun bind(studyMemberUIModel: StudyMemberUIModel) {
+        binding.studyMember = studyMemberUIModel
     }
 
     companion object {
